newapp/requirements.txt
python-dotenv
openai
Pillow
customtkinter